1) Improved delete entities system. 2) findAll's total annihilation (where it was needed). 3) Fixed car's validation; 4) Abstract things became more abstract. 5) Wraps added. 6) Consultant's rate fixed. 7) Builders for DTO's added. 8) General and cosmetic improves implemented. 9) Fresh JSON for Postman and DDL+DML added.

+++ src/main/java/carDealerApplication/entity/Consultant.java
@@ -5,6 +5,8 @@ import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode(callSuper = true)
 @Entity
 @Table(name = "consultant")
@@ -13,7 +15,8 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Consultant extends AbstractUserEntity {
-    private Double rate;
+    private BigDecimal rate;
+
     @OneToOne
     private Manufacturer manufacturer;
 }
